moved the strickly codegen related functions into csv_converter and left the actor code in csv2api.
