+++ csv_converter/Cargo.toml
@@ -4,6 +4,10 @@ version = "0.0.1"
 authors = ["Rob Rowe <robrowe04@gmail.com>"]
 
 [dependencies]
+barrel = "0.2.3"
+codegen = { git = "https://github.com/carllerche/codegen" } # Soon on crates.io
+csv = "1.0.0"
+futures = "0.1.23"
 serde = "1.0.70"
 serde_derive = "1.0.66"
 serde_json = "1.0.19"
