+++ csv_converter/src/lib.rs
@@ -1,6 +1,12 @@
+pub mod code_gen;
 pub mod config;
 pub mod input;
+pub mod models;
 
+extern crate barrel;
+extern crate codegen;
+extern crate csv;
+extern crate futures;
 #[macro_use]
 extern crate serde_derive;
